reset tests to master
