+++ Tests/Leader/Test_App_CommandLine.c
@@ -12,7 +12,7 @@ int main(void){
     BSP_UART_Init(UART_2);
     CANbus_Init(CAN_1, NULL, 0);
     Contactors_Init();
-    Minion_Init();
+    Minions_Init();
     Pedals_Init();
     Task_CommandLine(NULL);
 }
