+++ Tests/Leader/Test_App_Telemetry.c
@@ -1,10 +1,13 @@
 #include "os.h"
 #include "Tasks.h"
-#include "Minions.h"
 #include <bsp.h>
 #include "config.h"
 #include "common.h"
-#include "Contactors.h" 
+#include "Pedals.h"
+#include "Minions.h"
+#include "Contactors.h"
+#include "CANbus.h"
+#include "CAN_Queue.h"
 
 static OS_TCB Task1TCB;
 static CPU_STK Task1Stk[DEFAULT_STACK_SIZE];
@@ -13,24 +16,27 @@ void Task1(void *arg)
 {   
     CPU_Init();
     
-    BSP_UART_Init(UART_2);
-    Minion_Init();
+    CANbus_Init(CARCAN);
+    CAN_Queue_Init();
+    // BSP_UART_Init(UART_2);
+    Pedals_Init();
+    Minions_Init();
     Contactors_Init();
-
-    OS_ERR err;
 
     OS_CPU_SysTickInit(SystemCoreClock / (CPU_INT32U)OSCfg_TickRate_Hz);
 
-    // Initialize IgnitionContactor
+    OS_ERR err;
+
+    // Initialize Telemetry
     OSTaskCreate(
-        (OS_TCB*)&IgnCont_TCB,
-        (CPU_CHAR*)"IgnitionContactor",
-        (OS_TASK_PTR)Task_Contactor_Ignition,
+        (OS_TCB*)&Telemetry_TCB,
+        (CPU_CHAR*)"Telemetry",
+        (OS_TASK_PTR)Task_Telemetry,
         (void*)NULL,
-        (OS_PRIO)TASK_IGN_CONT_PRIO,
-        (CPU_STK*)IgnCont_Stk,
+        (OS_PRIO)TASK_TELEMETRY_PRIO,
+        (CPU_STK*)Telemetry_Stk,
         (CPU_STK_SIZE)WATERMARK_STACK_LIMIT,
-        (CPU_STK_SIZE)TASK_IGN_CONT_STACK_SIZE,
+        (CPU_STK_SIZE)TASK_TELEMETRY_STACK_SIZE,
         (OS_MSG_QTY)0,
         (OS_TICK)0,
         (void*)NULL,
@@ -39,12 +45,29 @@ void Task1(void *arg)
     );
     assertOSError(OS_MAIN_LOC, err);
 
+    bool lightState = false;
+    State contactorState = OFF;
+
+    CANDATA_t msg;
+    
     while (1){
-        printf("Array Contactor: %d, Motor Contactor: %d\n\r", Contactors_Get(ARRAY_CONTACTOR), Contactors_Get(MOTOR_CONTACTOR));
+        // Switches can be modified through hardware
+
+        // Check light
+        lightState = !lightState;
+        Minions_Write(BRAKELIGHT, lightState);
+
+        // Check contactors (HAVE NOTHING HOOKED UP TO CONTACTORS)
+        contactorState = contactorState == OFF ? ON : OFF;
+        Contactors_Set(MOTOR_CONTACTOR, contactorState, true);
+        Contactors_Set(ARRAY_CONTACTOR, contactorState, true);
+
+        CANbus_Read(&msg, CAN_BLOCKING, CARCAN);
 
         OSTimeDlyHMSM(0, 0, 0, 500, OS_OPT_TIME_HMSM_STRICT, &err);
+        // Use a logic analyzer to read the CAN line and see if the data shows up correctly    
     }
-};
+}
 
 int main()
 {
@@ -57,7 +80,7 @@ int main()
         (CPU_CHAR *)"Task 1",
         (OS_TASK_PTR)Task1,
         (void *)NULL,
-        (OS_PRIO)13,
+        (OS_PRIO)5,
         (CPU_STK *)Task1Stk,
         (CPU_STK_SIZE)DEFAULT_STACK_SIZE / 10,
         (CPU_STK_SIZE)DEFAULT_STACK_SIZE,
