+++ Tests/Leader/Test_Evac.c
@@ -1,0 +1,122 @@
+/* Copyright (c) 2020 UT Longhorn Racing Solar */
+
+#include "Tasks.h"
+#include "CANbus.h"
+#include "CAN_Queue.h"
+#include "ReadCarCAN.h"
+#include "Display.h"
+#include "UpdateDisplay.h"
+#include "CANConfig.h"
+
+/*
+ * NOTE: This test must be run with car CAN in loopback mode
+ * TODO: automate this, either with arguments to BSP or #define
+*/
+
+static OS_TCB Task1_TCB;
+#define STACK_SIZE 128
+static CPU_STK Task1_Stk[STACK_SIZE];
+
+#define CARCAN_FILTER_SIZE (sizeof carCANFilterList / sizeof(CANId_t))
+
+void Task1(){
+    OS_ERR err;
+
+    CPU_Init();
+    OS_CPU_SysTickInit(SystemCoreClock / (CPU_INT32U) OSCfg_TickRate_Hz);
+    CANbus_Init(CARCAN, carCANFilterList, CARCAN_FILTER_SIZE);
+    Contactors_Init();
+    Display_Init();
+    UpdateDisplay_Init();
+
+    // Send a bunch of charge enable messages to ReadCarCAN for 10 seconds
+    CANDATA_t data = {.ID = CHARGE_ENABLE, .idx = 0, .data = {1}};
+    for (int i=0; i<40; i++) {
+        CANbus_Send(data, true, CARCAN);
+        OSTimeDlyHMSM(0, 0, 0, 250, OS_OPT_TIME_HMSM_STRICT, &err);
+    }
+    data.ID = BPS_TRIP;
+    data.data[0] = 1;
+    CANbus_Send(data, true, CARCAN);
+    OSTimeDlyHMSM(1, 0, 0, 0, OS_OPT_TIME_HMSM_STRICT, &err);
+    while (1);
+}
+
+int main(){
+    OS_ERR err;
+    OSInit(&err);
+    assertOSError(OS_MAIN_LOC, err);
+
+    OSSemCreate(&FaultState_Sem4, "Fault State Semaphore", 0, &err);
+    OSTaskCreate(
+        (OS_TCB*)&Task1_TCB,
+        (CPU_CHAR*)"Task1",
+        (OS_TASK_PTR)Task1,
+        (void*)NULL,
+        (OS_PRIO)4,
+        (CPU_STK*)Task1_Stk,
+        (CPU_STK_SIZE)STACK_SIZE/10,
+        (CPU_STK_SIZE)STACK_SIZE,
+        (OS_MSG_QTY)0,
+        (OS_TICK)NULL,
+        (void*)NULL,
+        (OS_OPT)(OS_OPT_TASK_STK_CLR|OS_OPT_TASK_STK_CHK),
+        (OS_ERR*)&err
+    );
+    assertOSError(OS_MAIN_LOC, err);
+
+    OSTaskCreate(
+        (OS_TCB*)&UpdateDisplay_TCB,
+        (CPU_CHAR*)"Update Display",
+        (OS_TASK_PTR)Task_UpdateDisplay,
+        (void*)NULL,
+        (OS_PRIO)TASK_UPDATE_DISPLAY_PRIO,
+        (CPU_STK*)UpdateDisplay_Stk,
+        (CPU_STK_SIZE)STACK_SIZE/10,
+        (CPU_STK_SIZE)STACK_SIZE,
+        (OS_MSG_QTY)0,
+        (OS_TICK)NULL,
+        (void*)NULL,
+        (OS_OPT)(OS_OPT_TASK_STK_CLR|OS_OPT_TASK_STK_CHK),
+        (OS_ERR*)&err
+    );
+    assertOSError(OS_MAIN_LOC, err);
+
+    OSTaskCreate(
+        (OS_TCB*)&FaultState_TCB,
+        (CPU_CHAR*)"Fault State",
+        (OS_TASK_PTR)Task_FaultState,
+        (void*)NULL,
+        (OS_PRIO)TASK_FAULT_STATE_PRIO,
+        (CPU_STK*)FaultState_Stk,
+        (CPU_STK_SIZE)STACK_SIZE/10,
+        (CPU_STK_SIZE)STACK_SIZE,
+        (OS_MSG_QTY)0,
+        (OS_TICK)NULL,
+        (void*)NULL,
+        (OS_OPT)(OS_OPT_TASK_STK_CLR|OS_OPT_TASK_STK_CHK),
+        (OS_ERR*)&err
+    );
+    assertOSError(OS_MAIN_LOC, err);
+
+    OSTaskCreate(
+        (OS_TCB*)&ReadCarCAN_TCB,
+        (CPU_CHAR*)"ReadCarCAN",
+        (OS_TASK_PTR)Task_ReadCarCAN,
+        (void*)NULL,
+        (OS_PRIO)TASK_READ_CAR_CAN_PRIO,
+        (CPU_STK*)ReadCarCAN_Stk,
+        (CPU_STK_SIZE)STACK_SIZE/10,
+        (CPU_STK_SIZE)STACK_SIZE,
+        (OS_MSG_QTY)0,
+        (OS_TICK)NULL,
+        (void*)NULL,
+        (OS_OPT)(OS_OPT_TASK_STK_CLR|OS_OPT_TASK_STK_CHK),
+        (OS_ERR*)&err
+    );
+    assertOSError(OS_MAIN_LOC, err);
+
+    OSStart(&err);
+    assertOSError(OS_MAIN_LOC, err);
+}
+
